Create class IPhones, generate override methods.
