+++ src/homework/hw19/IOS.java
@@ -1,0 +1,5 @@
+package homework.hw19;
+
+public interface IOS {
+    void airplaneMode();
+}
