+++ src/homework/hw19/SmartphoneRunner.java
@@ -3,10 +3,10 @@ package homework.hw19;
 public class SmartphoneRunner {
     public static void main(String[] args) {
         Smartphones[] smartphones = {
-                new Android(),
-                new Android(),
-                new IPhone(),
-                new Iphone(),
+                new Androids(),
+                new Androids(),
+                new IPhones(),
+                new IPhones(),
         };
         long[] phoneNumber = {
                 380502582500L,
