--- src/homework/hw19/iOS.java
@@ -1,5 +1,0 @@
-package homework.hw19;
-
-public interface iOS {
-    void airplaneMod();
-}
